backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
python-dotenv==1.2.1
pgvector==0.4.2
psycopg2-binary==2.9.11
openai==2.21.0
httpx==0.28.1
feedparser==6.0.12
beautifulsoup4==4.14.3
markdownify==1.2.2
newspaper3k==0.2.8
lxml==6.0.2
lxml_html_clean==0.4.3
playwright==1.58.0
